create GoodsRegistration
